+++ tests/test_downloader_middlewares.py
@@ -159,18 +159,18 @@ class TestUserAgentMiddleware:
     def test_static_user_agent(self):
         user_agent = 'test user agent'
         mw = UserAgentMiddleware.from_crawler(Crawler(user_agent=user_agent))
-        req = HttpRequest('http://example.com')
+        req = HttpRequest('http://example.com', headers={})
         mw.handle_request(req)
         assert req.headers.get('User-Agent') == user_agent
 
     def test_gen_user_agent(self):
         mw = UserAgentMiddleware.from_crawler(Crawler(user_agent=':desktop,chrome'))
-        req = HttpRequest('http://example.com')
+        req = HttpRequest('http://example.com', headers={})
         mw.handle_request(req)
         assert 'Chrome' in req.headers.get('User-Agent')
 
         mw2 = UserAgentMiddleware.from_crawler(Crawler(user_agent=':mobile,chrome'))
-        req2 = HttpRequest('http://example.com')
+        req2 = HttpRequest('http://example.com', headers={})
         mw2.handle_request(req2)
         assert 'CriOS' in req2.headers.get('User-Agent') and 'Mobile' in req2.headers.get('User-Agent')
 
@@ -180,8 +180,8 @@ class TestUserAgentMiddleware:
 
     def test_random_user_agent(self):
         mw = UserAgentMiddleware.from_crawler(Crawler(random_user_agent=True))
-        req = HttpRequest('http://example.com')
-        req2 = HttpRequest('http://example.com')
+        req = HttpRequest('http://example.com', headers={})
+        req2 = HttpRequest('http://example.com', headers={})
         mw.handle_request(req)
         mw.handle_request(req2)
         assert 'User-Agent' in req.headers
@@ -191,7 +191,7 @@ class TestUserAgentMiddleware:
     def test_random_user_agent2(self):
         mw = UserAgentMiddleware.from_crawler(Crawler(user_agent=':mobile', random_user_agent=True))
         for i in range(30):
-            req = HttpRequest('http://example.com')
+            req = HttpRequest('http://example.com', headers={})
             mw.handle_request(req)
             assert 'User-Agent' in req.headers
             assert 'CriOS' in req.headers.get('User-Agent') and 'Mobile' in req.headers.get('User-Agent')